Completed exercise 4.21, delete blog operation only possible by authorized user

+++ part4/bloglist/controllers/blogs.js
@@ -44,8 +44,19 @@ blogRouter.post('/', async (request, response) => {
 
 
 blogRouter.delete('/:id', async (request, response) => {
-	await Blog.findByIdAndDelete(request.params.id)
-	response.status(204).end()
+	const decodedToken = jwt.verify(request.token, process.env.SECRET)
+	if (!decodedToken.id) {
+		return response.status(401).json({ error: 'token invalid' })
+	}
+	const user = await User.findById(decodedToken.id)
+	const blog = await Blog.findById(request.params.id)
+
+	if (blog.user.toString() === user.id.toString()) {
+		await Blog.findByIdAndDelete(request.params.id)
+		response.status(204).end()
+	} else {
+		response.status(401).json({ error: 'unauthorized' })
+	}
 })
 
 blogRouter.put('/:id', async (request, response) => {
